+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Add support for nobrainer, an ORM for RethinkDB
 - Validation error message on invalid registration form submission
 
 ## [3.0.2] - 2021-02-14